[File] add Group Model , File : User ( email , name , password, ProfileImage
