+++ BlanketKick/Model/User.swift
@@ -1,0 +1,9 @@
+import Foundation
+import SwiftUI
+
+struct User {
+    var email: String
+    var name: String
+    var password: String
+    var profileImage: UIImage?
+}
